+++ src/app/(main)/settings/llm/ProviderList/providers.tsx
@@ -11,6 +11,7 @@ import {
   Minimax,
   Mistral,
   Moonshot,
+  Novita,
   OpenRouter,
   Perplexity,
   Stepfun,
@@ -35,6 +36,7 @@ import {
   MinimaxProviderCard,
   MistralProviderCard,
   MoonshotProviderCard,
+  NovitaProviderCard,
   OpenRouterProviderCard,
   PerplexityProviderCard,
   QwenProviderCard,
@@ -131,6 +133,11 @@ export const useProviderList = (): ProviderItem[] => {
         ...OpenRouterProviderCard,
         docUrl: urlJoin(BASE_DOC_URL, 'openrouter'),
         title: <OpenRouter.Combine iconProps={{ color: OpenRouter.colorPrimary }} size={20} />,
+      },
+      {
+        ...NovitaProviderCard,
+        docUrl: urlJoin(BASE_DOC_URL, 'novita'),
+        title: <Novita.Combine size={20} type={'color'} />,
       },
       {
         ...TogetherAIProviderCard,