+++ src/components/ModelProviderIcon/index.tsx
@@ -12,6 +12,7 @@ import {
   Minimax,
   Mistral,
   Moonshot,
+  Novita,
   Ollama,
   OpenAI,
   OpenRouter,
@@ -115,6 +116,10 @@ const ModelProviderIcon = memo<ModelProviderIconProps>(({ provider }) => {
 
     case ModelProvider.Stepfun: {
       return <Stepfun size={20} />;
+    }
+
+    case ModelProvider.Novita: {
+      return <Novita size={20} />;
     }
 
     case ModelProvider.Baichuan: {